feat: 알람 설정 테이블 구현 및 타입 정리
