+++ src/shared/component/Table/Table.stories.tsx
@@ -1,8 +1,8 @@
-import { StudyListData, alarmSettingsData } from "@/shared/constant/example";
+import { AlarmSettingsData, StudyListData } from "@/shared/constant/example";
 import { theme } from "@/styles/themes.css";
 
+import { ALARM_SETTINGS_COLUMNS } from "@/view/user/setting/AlarmSetting/AlarmSettingTable/constant";
 import { STUDY_LIST_COLUMNS } from "@/view/user/setting/StudyList/StudyListTable/constant";
-import { ALARM_SETTINGS_COLUMNS } from "@/view/user/setting/constant";
 import type { Meta } from "@storybook/react";
 import { DataTable } from ".";
 
@@ -42,7 +42,7 @@ export const AlarmSetting = {
     return (
       <DataTable
         type="알림설정"
-        rows={alarmSettingsData}
+        rows={AlarmSettingsData}
         cols={ALARM_SETTINGS_COLUMNS}
       />
     );
