+++ src/view/user/setting/AlarmSetting/AlarmSettingTable/type.ts
@@ -1,0 +1,9 @@
+export type AlarmSettingsDataType = {
+  alarm: boolean;
+  groupName: string;
+  problemRegistration: boolean;
+  solutionRegistration: boolean;
+  commentRegistration: boolean;
+  newMemberAllowed: boolean;
+  endDateImminent: boolean;
+};
