+++ src/view/user/setting/AlarmSetting/index.tsx
@@ -1,0 +1,13 @@
+import { headingStyle, wrapperStyle } from "../index.css";
+import AlarmSettingTable from "./AlarmSettingTable";
+
+const AlarmSetting = () => {
+  return (
+    <article className={wrapperStyle({ type: "알림설정" })}>
+      <h1 className={headingStyle}>알람 설정</h1>
+      <AlarmSettingTable />
+    </article>
+  );
+};
+
+export default AlarmSetting;
